rf: ZENKO-744 rename lifecycle producer and consumer files
