+++ extensions/lifecycle/objectProcessor/LifecycleObjectProcessor.js
