+++ package.json
@@ -9,8 +9,8 @@
     "ingestion_populator": "node bin/ingestion.js",
     "replication_status_processor": "node extensions/replication/replicationStatusProcessor/task.js",
     "lifecycle_conductor": "node extensions/lifecycle/conductor/service.js",
-    "lifecycle_producer": "node extensions/lifecycle/lifecycleProducer/task.js",
-    "lifecycle_consumer": "node extensions/lifecycle/lifecycleConsumer/task.js",
+    "lifecycle_producer": "node extensions/lifecycle/bucketProcessor/task.js",
+    "lifecycle_consumer": "node extensions/lifecycle/objectProcessor/task.js",
     "mongo_queue_processor": "node extensions/mongoProcessor/mongoProcessorTask.js",
     "garbage_collector": "node extensions/gc/service.js",
     "test": "mocha --recursive tests/unit",
